move auth resolvers
